feat(@dpc-sdp/ripple-tide-api): first pass at global element refactor
