+++ examples/nuxt-app/error.vue
@@ -1,0 +1,1 @@
+<template>500 error here</template>
